Renaming, redoing styles here and there
