+++ app/router.js
@@ -167,7 +167,7 @@ Router.map(function() {
       });
       this.route('setup-complete');
     });
-    this.route('visited-first-steps');
+    this.route('first-steps');
   });
 
   this.route('error');
