+++ app/styles/granite/page/first-steps.less
@@ -1,0 +1,120 @@
+.ui.hero.cover.hero__first-steps-cover {
+  margin-top: -20px;
+  min-height: 90vh;
+}
+
+.first-steps {
+  h1.ui.header.first-steps__header {
+    font-size: 3em;
+    margin-bottom: 1em;
+    opacity: .9;
+
+    > small {
+      font-size: .5em;
+    }
+  }
+
+  .first-steps__card {
+    padding: 1.5em 2em;
+
+    &.first-steps__card--completed {
+      .three-point-gradient(fade(@granite-green, 80%), fade(lighten(@granite-green, 5%), 80%), fade(darken(@granite-green, 10%), 80%));
+      background: @gradient;
+      background-color: @granite-grey;
+      background-repeat: no-repeat;
+    }
+  }
+}
+
+.first-steps .segment > h3 {
+  font-size: 2.5em;
+}
+
+.first-steps .segment > .spacing + .label > p{
+  font-size: 1.2em;
+}
+
+.first-steps .segment > p {
+  font-size: 1.2em;
+  margin-bottom:77px;
+}
+
+// .first-steps .column {
+//   margin: .8em;
+//   margin-top: 6em;
+// }
+
+.first-steps .column > .segment {
+    margin: .8em;
+    display: block;
+}
+
+.first-steps .segment > .spacing + .label > .icon > p{
+  font-size: 1.6em;
+  margin-top: 1.5em;
+}
+
+// .first-steps .grid {
+//   padding: 10px;
+// }
+
+//----AFTER COMPLETED----//
+.first-steps .segment > .done__icon > i{
+  color: @gray-lighter;
+  position: absolute;
+  display: block;
+  margin-left: 0.7em;
+  margin-top: 0.4em;
+}
+
+.first-steps .segment > .spacing + .label > .overlay__small > p{
+  font-size: 1.3em;
+}
+
+
+
+.first-steps .first-steps__steps i:first-child {
+  color: @granite-green;
+  position: absolute;
+  display: block;
+  margin-left: 14.7em;
+  margin-top: -1.4em;
+  z-index: 2;
+}
+
+.first-steps .segment > .overlay {
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  z-index: 2;
+}
+
+.first-steps .segment > .label > .done__icon__location{
+  position: absolute;
+  // display: none;
+  width: 100%;
+  height: 50%;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(102,217,100,0.8);
+  z-index: 2;
+}
+
+.first-steps .segment > .label > .done__icon__department{
+  position: absolute;
+  // display: none;
+  width: 100%;
+  height: 50%;
+  top: 3.6em;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  background-color: rgba(102,217,100,0.8);
+  z-index: 2;
+}
