+++ app/styles/granite/ui/heros.less
@@ -10,6 +10,7 @@
 
   .makeHeroImage(~'granite-cover');
   .makeHeroImage(~'lift-cover');
+  .makeHeroImage(~'first-steps-cover');
 
   &.grey.overlay {
     background-color: @granite-grey;
